# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebBanSachOnline API",
    "version": "1.0.0",
    "description": "Tài liệu API cho dự án WebBanSachOnline"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Đăng ký tài khoản mới",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "full_name",
                  "phone_number",
                  "email",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "full_name": {
                    "type": "string"
                  },
                  "phone_number": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Thành công"
          },
          "400": {
            "description": "Lỗi validate"
          },
          "409": {
            "description": "Trùng lặp dữ liệu"
          }
        }
      }
    },
    "/auth/resolve-identifier": {
      "post": {
        "summary": "Tra cứu email dựa trên tên đăng nhập",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "identifier"
                ],
                "properties": {
                  "identifier": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trả về email hợp lệ"
          },
          "400": {
            "description": "Thiếu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          }
        }
      }
    },
...
  }
}